Move StatusOr to common package
