+++ fe/fe-core/src/main/java/com/starrocks/load/batchwrite/RequestCoordinatorBackendResult.java
@@ -14,6 +14,7 @@
 
 package com.starrocks.load.batchwrite;
 
+import com.starrocks.common.StatusOr;
 import com.starrocks.system.ComputeNode;
 import com.starrocks.thrift.TStatus;
